+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## [2.0.0] - 2024-03-20
+
+- Added arguments for `Mailtrap::Client`
+  - `bulk` to use Mailtrap bulk sending API
+  - `sandbox` to use Mailtrap sandbox API
+  - `inbox_id` required when using Mailtrap sandbox API
+
+- Removed Sending namespace, affected classes:
+  - `Mailtrap::Sending::Client` -> `Mailtrap::Client`
+  - `Mailtrap::Sending::Error` -> `Mailtrap::Error`
+  - `Mailtrap::Sending::AttachmentContentError` -> `Mailtrap::AttachmentContentError`
+  - `Mailtrap::Sending::AuthorizationError` -> `Mailtrap::AuthorizationError`
+  - `Mailtrap::Sending::MailSizeError` -> `Mailtrap::MailSizeError`
+  - `Mailtrap::Sending::RateLimitError` -> `Mailtrap::RateLimitError`
+  - `Mailtrap::Sending::RejectionError` -> `Mailtrap::RejectionError`
+
 ## [1.2.2] - 2023-11-01
 
 - Improved error handling